+++ src/Naming.cpp
@@ -443,7 +443,7 @@ Utf8 Naming::aKindName(ScopeKind kind)
     return article;
 }
 
-Utf8 Naming::niceArgumentRank(int idx)
+Utf8 Naming::niceArgumentRank(size_t idx)
 {
     switch (idx)
     {
@@ -457,7 +457,7 @@ Utf8 Naming::niceArgumentRank(int idx)
     return Fmt("argument '%d'", idx);
 }
 
-Utf8 Naming::niceParameterRank(int idx)
+Utf8 Naming::niceParameterRank(size_t idx)
 {
     switch (idx)
     {
