+++ src/SemanticJob_Op.cpp
@@ -278,21 +278,21 @@ bool SemanticJob::checkFuncPrototypeOp(SemanticContext* context, AstFuncDecl* no
     {
         SWAG_CHECK(checkFuncPrototypeOpNumParams(context, node, parameters, 2, false));
         SWAG_CHECK(checkFuncPrototypeOpReturnType(context, node, nullptr));
-        for (size_t i = 1; i < parameters->childs.size(); i++)
+        for (uint32_t i = 1; i < (uint32_t) parameters->childs.size(); i++)
             SWAG_CHECK(checkFuncPrototypeOpParam(context, node, parameters, i, g_TypeMgr->typeInfoU64));
     }
     else if (name == g_LangSpec->name_opIndexAssign)
     {
         SWAG_CHECK(checkFuncPrototypeOpNumParams(context, node, parameters, 3, false));
         SWAG_CHECK(checkFuncPrototypeOpReturnType(context, node, g_TypeMgr->typeInfoVoid));
-        for (size_t i = 1; i < parameters->childs.size() - 1; i++)
+        for (uint32_t i = 1; i < (uint32_t) parameters->childs.size() - 1; i++)
             SWAG_CHECK(checkFuncPrototypeOpParam(context, node, parameters, i, g_TypeMgr->typeInfoU64));
     }
     else if (name == g_LangSpec->name_opIndexAffect)
     {
         SWAG_CHECK(checkFuncPrototypeOpNumParams(context, node, parameters, 3, false));
         SWAG_CHECK(checkFuncPrototypeOpReturnType(context, node, g_TypeMgr->typeInfoVoid));
-        for (size_t i = 1; i < parameters->childs.size() - 1; i++)
+        for (uint32_t i = 1; i < (uint32_t) parameters->childs.size() - 1; i++)
             SWAG_CHECK(checkFuncPrototypeOpParam(context, node, parameters, i, g_TypeMgr->typeInfoU64));
     }
     else if (name == g_LangSpec->name_opInit && node->sourceFile->isGenerated)
